Sixth attempt to fix SonarCloud configuration
It turns out that the *JavaScript* exclusion filter kicks all Java source code out of the list of files to be analyzed. This occurs only in builds made on Travis CI, and not in Sonar analyses performed locally.

Attempting to override the exclusion list for JS and TS by setting their parameter to the empty string.