Change to the contract of SynchronousProcessor.compute
This commit deserves some explanation as it slightly changes the behavior of SynchronousProcessor. In its original implementation, a call to compute() could not at the same time put events in the output queue and return false. In such a case, the events in the output queue were ignored (i.e. not propagated downstream). Thus the semantics of a false return value were "I will never produce output events and this starts now".

The new implementation of Pushable and Pullable in this class do let the output events through; the new semantics is thus "I will never produce output events on *future calls* to compute".