Support for Pushable.pushFast is dropped
This method was introduced around 2016-2017, and was intended to offer the possibility of pushing events in a non-blocking way. Thus one could push events in rapid-fire fashion to processors in distinct threads, allowing them to run in parallel, and then collect their results once they were done.

Successfully achieving performance gains required much fine-tuning, and could be done for a subset of all possible processors. For most processors, pushFast simply contained a call to push. No project that we know of makes use of these functionalities, and it was decided to simplify the code by removing this option.

As an upside, the removal of this non-blocking mode might provide a slight performance boost for *blocking* (i.e. normal) mode, due to a simpler management in some processors, such as Slice.