Fix to GitHub Action workflow
Now that there is a pom.xml file in the project's root, the build action
fails with the message "Maven project detected. You should run the goal
'org.sonarsource.scanner.maven:sonar' during build rather than using
this GitHub Action."

Yet, pom.xml is there only for documentation, it is not a Maven project.
The fix is there:

https://community.sonarsource.com/t/maven-project-detected-you-should-run-the-goal-org-sonarsource-scanner-maven-sonar-during-build-rather-than-using-this-github-action/36158/1

One must delete pom.xml in one of the actions.