Fix to ReadInputStream
When reading from named pipes, it would eventually stop if no new bytes were read after some time. This was caused by the fact that SynchronousProcessor has a *bounded* loop in OutputPullable.hasNext, which halts after a large but finite number of attempts. The fix is to have ReadInputStream provide a different Pullable object that waits forever.