Removed dependency to Petit Poucet 0.2
This library was used to store relationships between input and output events
of a processor, thus giving a primitive way of obtaining provenance information.
However, Petit Poucet has been refactored and rewritten many times, and BeepBeep
still carried this outdated and quirky version.

The first step to an upgrade is to purely and simply remove anything related to
provenance and drop the dependency on the library. This is what this release
does. Note that this does not break the API: methods having an EventTracker as
an argument are still there, marked as deprecated, and act as if no tracker is
passed.

As a result, the code is simpler: about 800 lines have been removed from the
core library, and 1,000 in the palettes. The compiled JAR has also been reduced
by about 10 kb compared to the previous version.